chore(pg_srv): Introduce separate crate for PostgreSQL server emulation
