+++ rust/cubesql/cubesql/src/sql/postgres/mod.rs
@@ -1,7 +1,5 @@
-pub(crate) mod buffer;
 pub(crate) mod extended;
 pub(crate) mod pg_type;
-pub(crate) mod protocol;
 pub(crate) mod service;
 pub(crate) mod shim;
 pub(crate) mod writer;
