+++ rust/cubesql/pg-srv/README.md
@@ -1,0 +1,16 @@
+<p align="center"><a href="https://cube.dev"><img src="https://i.imgur.com/zYHXm4o.png" alt="Cube.js" width="300px"></a></p>
+
+[Website](https://cube.dev) • [Docs](https://cube.dev/docs) •
+[Examples](#examples) • [Blog](https://cube.dev/blog) •
+[Slack](https://slack.cube.dev) • [Twitter](https://twitter.com/thecubejs)
+
+[![GitHub Actions](https://github.com/cube-js/cube.js/workflows/Rust/badge.svg)](https://github.com/cube-js/cube.js/actions?query=workflow%3ARust+branch%3Amaster)
+[![FOSSA Status](https://app.fossa.io/api/projects/git%2Bgithub.com%2Fcube-js%2Fcube.js.svg?type=shield)](https://app.fossa.io/projects/git%2Bgithub.com%2Fcube-js%2Fcube.js?ref=badge_shield)
+
+# Cube SQL API
+
+Library for emulating a PostgreSQL server.
+
+## License
+
+Cube Store is [Apache 2.0 licensed](./LICENSE).
