+++ rust/cubesql/pg-srv/src/buffer.rs
