+++ rust/cubesql/pg-srv/src/extended.rs
@@ -1,0 +1,10 @@
+#[derive(Debug)]
+pub enum BindValue {
+    String(String),
+    Int64(i64),
+    #[allow(unused)]
+    UInt64(u64),
+    Float64(f64),
+    Bool(bool),
+    Null,
+}
